v0.0.1-beta

+++ ClienteApp/Helpers/PSHelper.cs
@@ -17,7 +17,7 @@ namespace ClienteApp.Helpers
             try
             {
                 process.Start();
-                string nombreEquipo = process.StandardOutput.ReadToEnd().Trim(); // Eliminar espacios en blanco
+                string nombreEquipo = process.StandardOutput.ReadToEnd().Trim();
                 string error = process.StandardError.ReadToEnd();
 
                 process.WaitForExit();
@@ -51,7 +51,9 @@ namespace ClienteApp.Helpers
             //Unistall Apps
             foreach (Models.App app in existingData?.grupo?.apps ?? [])
             {
+                Console.ForegroundColor = ConsoleColor.Red;
                 Console.WriteLine($"Procesando aplicación para desinstalación: {app.name} (Versión: {app.version})");
+                Console.ResetColor();
                 bool isDifferentGroup = existingData != null && existingData.grupo != null && existingData.grupo.name != data.grupo.name;
                 bool isAppNotInNewData = data.grupo.apps.All(newApp =>
                     newApp.name != app.name ||
@@ -59,14 +61,16 @@ namespace ClienteApp.Helpers
 
                 if (isDifferentGroup || isAppNotInNewData)
                 {
-                    DesinstalarAplicacion(app);
+                    DesInstalarAplicacion(app);
                 }
             }
 
             // Install Apps
             foreach (Models.App app in data.grupo.apps)
             {
+                Console.ForegroundColor = ConsoleColor.Green;
                 Console.WriteLine($"Procesando aplicación para instalación: {app.name} (Versión: {app.version})");
+                Console.ResetColor();
                 bool install = true;
                 bool isAppAlreadyInstalled = existingData != null && existingData.grupo != null && existingData.grupo.name == data.grupo.name &&
                                              existingData.grupo.apps != null && existingData.grupo.apps.Any(existingApp => existingApp.name == app.name && existingApp.version == app.version) == true;
@@ -90,12 +94,24 @@ namespace ClienteApp.Helpers
                 return;
             }
 
-            string tempFilePath = Path.Combine(Path.GetTempPath(), "ClientApp", Path.GetFileName(app.path));
-            Console.WriteLine($"Carpeta temporal: {tempFilePath}");
-
+            string tempFolder = Path.Combine(Path.GetTempPath(), "ClientApp");
+            string tempFilePath = Path.Combine(tempFolder, Path.GetFileName(app.path));
             try
             {
-                File.Copy(app.path, tempFilePath, true);
+                if (!Directory.Exists(tempFolder))
+                {
+                    Directory.CreateDirectory(tempFolder);
+                }
+
+                if (!File.Exists(tempFilePath))
+                {
+                    Console.WriteLine($"Archivo no encontrado en la carpeta temporal. Copiando...");
+                    File.Copy(app.path, tempFilePath, true);
+                }
+                else
+                {
+                    Console.WriteLine($"Archivo encontrado en la carpeta temporal: {tempFilePath}");
+                }
 
                 using Process process = new();
                 process.StartInfo.FileName = "msiexec";
@@ -127,51 +143,51 @@ namespace ClienteApp.Helpers
             {
                 Console.WriteLine($"Excepción al instalar la aplicación {app.name}: {ex.Message}");
             }
-            finally
-            {
-                // Intentar eliminar el archivo temporal después de la instalación
-                try
-                {
-                    if (File.Exists(tempFilePath))
-                    {
-                        File.Delete(tempFilePath);
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Error al eliminar el archivo temporal: {ex.Message}");
-                }
-            }
         }
 
-        private static void DesinstalarAplicacion(Models.App app)
+        private static void DesInstalarAplicacion(Models.App app)
         {
             if (string.IsNullOrEmpty(app.name) || string.IsNullOrEmpty(app.version) || string.IsNullOrEmpty(app.path))
             {
                 return;
             }
-            string tempFilePath = Path.Combine(Path.GetTempPath(), Path.GetFileName(app.path));
+
+            string tempFolder = Path.Combine(Path.GetTempPath(), "ClientApp");
+            string tempFilePath = Path.Combine(tempFolder, Path.GetFileName(app.path));
+
             try
             {
-                // Copiar el archivo desde la carpeta compartida a la ubicación temporal
-                File.Copy(app.path, tempFilePath, true);
+                if (!Directory.Exists(tempFolder))
+                {
+                    Directory.CreateDirectory(tempFolder);
+                }
 
-                using Process process = new();
-                process.StartInfo.FileName = "msiexec";
-                process.StartInfo.Arguments = $"/x \"{tempFilePath}\" /quiet /norestart";
-                process.StartInfo.RedirectStandardOutput = true;
-                process.StartInfo.RedirectStandardError = true;
-                process.StartInfo.UseShellExecute = false; // Para redirigir la salida
-                process.StartInfo.CreateNoWindow = true;   // No mostrar ventanas
-                process.StartInfo.Verb = "runas";         // Ejecutar como administrador
+                if (!File.Exists(tempFilePath))
+                {
+                    Console.WriteLine($"Archivo no encontrado en la carpeta temporal. Copiando desde:");
+                    File.Copy(app.path, tempFilePath, true);
+                }
+                else
+                {
+                    Console.WriteLine($"Archivo encontrado en la carpeta temporal: {tempFilePath}");
+                }
+
+
+                using Process process2 = new();
+                process2.StartInfo.FileName = "msiexec";
+                process2.StartInfo.Arguments = $"/x \"{tempFilePath}\" /quiet /norestart";
+                process2.StartInfo.RedirectStandardOutput = true;
+                process2.StartInfo.RedirectStandardError = true;
+                process2.StartInfo.UseShellExecute = false; // Para redirigir la salida
+                process2.StartInfo.Verb = "runas";         // Ejecutar como administrador
 
                 Console.WriteLine($"Desinstalando la aplicación {app.name}...");
-                process.Start();
+                process2.Start();
 
-                string output = process.StandardOutput.ReadToEnd();
-                string error = process.StandardError.ReadToEnd();
+                string output = process2.StandardOutput.ReadToEnd();
+                string error = process2.StandardError.ReadToEnd();
 
-                process.WaitForExit();
+                process2.WaitForExit();
 
                 if (!string.IsNullOrWhiteSpace(error))
                 {
@@ -185,21 +201,6 @@ namespace ClienteApp.Helpers
             catch (Exception ex)
             {
                 Console.WriteLine($"Excepción al desinstalar la aplicación {app.name}: {ex.Message}");
-            }
-            finally
-            {
-                try
-                {
-                    if (File.Exists(tempFilePath))
-                    {
-                        File.Delete(tempFilePath);
-                        Console.WriteLine("Archivo temporal eliminado correctamente.");
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Error al eliminar el archivo temporal: {ex.Message}");
-                }
             }
         }
     }
